Splits the demos in different Activities
